add it data dummy for tbVists

+++ database/migrations/2023_05_16_211655_create_feed_backs_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('id_visit')->constrained('visits');
             $table->text('text_feed');
-            $table->integer('skala_feed');
+            $table->unsignedTinyInteger('skala_feed'); // Menggunakan tipe data unsignedTinyInteger
             $table->timestamps();
         });
     }
